[Refactor] : train.py 및 학습진행
